+++ DependencyInjection/Configuration.php
@@ -11,11 +11,13 @@
 
 namespace Hackzilla\Bundle\TicketBundle\DependencyInjection;
 
+use Hackzilla\Bundle\TicketBundle\Entity\Ticket;
+use Hackzilla\Bundle\TicketBundle\Entity\TicketMessage;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
 /**
- * This is the class that validates and merges configuration from your app/config files.
+ * This is the class that validates and merges configuration from your config files.
  *
  * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html#cookbook-bundles-extension-config-class}
  *
@@ -29,22 +31,17 @@ class Configuration implements ConfigurationInterface
     public function getConfigTreeBuilder()
     {
         $treeBuilder = new TreeBuilder('hackzilla_ticket');
-        if (method_exists($treeBuilder, 'getRootNode')) {
-            $rootNode = $treeBuilder->getRootNode();
-        } else {
-            // BC layer for symfony/config 4.1 and older
-            $rootNode = $treeBuilder->root('hackzilla_ticket');
-        }
 
-        $rootNode
+        $treeBuilder
+            ->getRootNode()
             ->children()
                 ->enumNode('translation_domain')
                     ->values(['HackzillaTicketBundle', 'messages'])
                     ->defaultValue('messages')
                 ->end()
                 ->scalarNode('user_class')->isRequired()->cannotBeEmpty()->end()
-                ->scalarNode('ticket_class')->cannotBeEmpty()->defaultValue('Hackzilla\Bundle\TicketBundle\Entity\Ticket')->end()
-                ->scalarNode('message_class')->cannotBeEmpty()->defaultValue('Hackzilla\Bundle\TicketBundle\Entity\TicketMessage')->end()
+                ->scalarNode('ticket_class')->cannotBeEmpty()->defaultValue(Ticket::class)->end()
+                ->scalarNode('message_class')->cannotBeEmpty()->defaultValue(TicketMessage::class)->end()
                 ->arrayNode('features')
                     ->addDefaultsIfNotSet()
                     ->children()