+++ Tests/Form/Type/TicketMessageTypeTest.php
@@ -23,7 +23,7 @@ final class TicketMessageTypeTest extends TypeTestCase
 {
     private $user;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->user = $this->createMock(UserManagerInterface::class);